Modelleri genişletme

+++ site/panel/application/controllers/Brands.php
@@ -13,19 +13,20 @@ class Brands extends VS_Controller
 
         $this->load->model("brand_model");
 
-        if(!get_active_user()){
+        if (!get_active_user()) {
             redirect(base_url("login"));
         }
-
     }
 
-    public function index(){
+    public function index()
+    {
 
         $viewData = new stdClass();
 
         /** Tablodan Verilerin Getirilmesi.. */
         $items = $this->brand_model->get_all(
-            array(), "rank ASC"
+            array(),
+            "rank ASC"
         );
 
         /** View'e gönderilecek Değişkenlerin Set Edilmesi.. */
@@ -36,8 +37,11 @@ class Brands extends VS_Controller
         $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index", $viewData);
     }
 
-    public function new_form(){
-
+    public function new_form()
+    {
+        if (!isAllowedWriteViewModule()) {
+            redirect(base_url("brands"));
+        }
         $viewData = new stdClass();
 
         /** View'e gönderilecek Değişkenlerin Set Edilmesi.. */
@@ -45,16 +49,16 @@ class Brands extends VS_Controller
         $viewData->subViewFolder = "add";
 
         $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index", $viewData);
-
     }
 
-    public function save(){
+    public function save()
+    {
 
         $this->load->library("form_validation");
 
         // Kurallar yazilir..
 
-        if($_FILES["img_url"]["name"] == ""){
+        if ($_FILES["img_url"]["name"] == "") {
 
             $alert = array(
                 "title" => "İşlem Başarısız",
@@ -81,15 +85,15 @@ class Brands extends VS_Controller
         // Form Validation Calistirilir..
         $validate = $this->form_validation->run();
 
-        if($validate){
+        if ($validate) {
 
             // Upload Süreci...
 
             $file_name = convertToSEO(pathinfo($_FILES["img_url"]["name"], PATHINFO_FILENAME)) . "." . pathinfo($_FILES["img_url"]["name"], PATHINFO_EXTENSION);
 
-            $image_350x216 = upload_picture($_FILES["img_url"]["tmp_name"], "uploads/$this->viewFolder",350,216, $file_name);
+            $image_350x216 = upload_picture($_FILES["img_url"]["tmp_name"], "uploads/$this->viewFolder", 350, 216, $file_name);
 
-            if($image_350x216){
+            if ($image_350x216) {
 
                 $insert = $this->brand_model->add(
                     array(
@@ -102,14 +106,13 @@ class Brands extends VS_Controller
                 );
 
                 // TODO Alert sistemi eklenecek...
-                if($insert){
+                if ($insert) {
 
                     $alert = array(
                         "title" => "İşlem Başarılı",
                         "text" => "Kayıt başarılı bir şekilde eklendi",
                         "type"  => "success"
                     );
-
                 } else {
 
                     $alert = array(
@@ -118,7 +121,6 @@ class Brands extends VS_Controller
                         "type"  => "error"
                     );
                 }
-
             } else {
 
                 $alert = array(
@@ -132,14 +134,12 @@ class Brands extends VS_Controller
                 redirect(base_url("brands/new_form"));
 
                 die();
-
             }
 
             // İşlemin Sonucunu Session'a yazma işlemi...
             $this->session->set_flashdata("alert", $alert);
 
             redirect(base_url("brands"));
-
         } else {
 
             $viewData = new stdClass();
@@ -151,11 +151,13 @@ class Brands extends VS_Controller
 
             $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index", $viewData);
         }
-
     }
 
-    public function update_form($id){
-
+    public function update_form($id)
+    {
+        if (!isAllowedUpdateViewModule()) {
+            redirect(base_url("brands"));
+        }
         $viewData = new stdClass();
 
         /** Tablodan Verilerin Getirilmesi.. */
@@ -164,19 +166,17 @@ class Brands extends VS_Controller
                 "id"    => $id,
             )
         );
-        
+
         /** View'e gönderilecek Değişkenlerin Set Edilmesi.. */
         $viewData->viewFolder = $this->viewFolder;
         $viewData->subViewFolder = "update";
         $viewData->item = $item;
 
         $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index", $viewData);
-
-
     }
 
-
-    public function update($id){
+    public function update($id)
+    {
 
         $this->load->library("form_validation");
 
@@ -193,22 +193,21 @@ class Brands extends VS_Controller
         // Form Validation Calistirilir..
         $validate = $this->form_validation->run();
 
-        if($validate){
+        if ($validate) {
 
             // Upload Süreci...
-            if($_FILES["img_url"]["name"] !== "") {
+            if ($_FILES["img_url"]["name"] !== "") {
 
                 $file_name = convertToSEO(pathinfo($_FILES["img_url"]["name"], PATHINFO_FILENAME)) . "." . pathinfo($_FILES["img_url"]["name"], PATHINFO_EXTENSION);
 
-                $image_350x216 = upload_picture($_FILES["img_url"]["tmp_name"], "uploads/$this->viewFolder",350,216, $file_name);
+                $image_350x216 = upload_picture($_FILES["img_url"]["tmp_name"], "uploads/$this->viewFolder", 350, 216, $file_name);
 
-                if($image_350x216){
+                if ($image_350x216) {
 
                     $data = array(
                         "title" => $this->input->post("title"),
                         "img_url" => $file_name,
                     );
-
                 } else {
 
                     $alert = array(
@@ -222,28 +221,24 @@ class Brands extends VS_Controller
                     redirect(base_url("brands/update_form/$id"));
 
                     die();
-
                 }
-
             } else {
 
                 $data = array(
                     "title" => $this->input->post("title"),
                 );
-
             }
 
             $update = $this->brand_model->update(array("id" => $id), $data);
 
             // TODO Alert sistemi eklenecek...
-            if($update){
+            if ($update) {
 
                 $alert = array(
                     "title" => "İşlem Başarılı",
                     "text" => "Kayıt başarılı bir şekilde güncellendi",
                     "type"  => "success"
                 );
-
             } else {
 
                 $alert = array(
@@ -257,7 +252,6 @@ class Brands extends VS_Controller
             $this->session->set_flashdata("alert", $alert);
 
             redirect(base_url("brands"));
-
         } else {
 
             $viewData = new stdClass();
@@ -276,11 +270,13 @@ class Brands extends VS_Controller
 
             $this->load->view("{$viewData->viewFolder}/{$viewData->subViewFolder}/index", $viewData);
         }
-
     }
 
-    public function delete($id){
-
+    public function delete($id)
+    {
+        if (!isAllowedDeleteViewModule()) {
+            redirect(base_url("brands"));
+        }
         $delete = $this->brand_model->delete(
             array(
                 "id"    => $id
@@ -288,14 +284,13 @@ class Brands extends VS_Controller
         );
 
         // TODO Alert Sistemi Eklenecek...
-        if($delete){
+        if ($delete) {
 
             $alert = array(
                 "title" => "İşlem Başarılı",
                 "text" => "Kayıt başarılı bir şekilde silindi",
                 "type"  => "success"
             );
-
         } else {
 
             $alert = array(
@@ -303,19 +298,16 @@ class Brands extends VS_Controller
                 "text" => "Kayıt silme sırasında bir problem oluştu",
                 "type"  => "error"
             );
-
-
         }
 
         $this->session->set_flashdata("alert", $alert);
         redirect(base_url("brands"));
-
-
     }
 
-    public function isActiveSetter($id){
+    public function isActiveSetter($id)
+    {
 
-        if($id){
+        if ($id) {
 
             $isActive = ($this->input->post("data") === "true") ? 1 : 0;
 
@@ -330,7 +322,8 @@ class Brands extends VS_Controller
         }
     }
 
-    public function rankSetter(){
+    public function rankSetter()
+    {
 
 
         $data = $this->input->post("data");
@@ -339,7 +332,7 @@ class Brands extends VS_Controller
 
         $items = $order["ord"];
 
-        foreach ($items as $rank => $id){
+        foreach ($items as $rank => $id) {
 
             $this->brand_model->update(
                 array(
@@ -350,9 +343,6 @@ class Brands extends VS_Controller
                     "rank"      => $rank
                 )
             );
-
         }
-
     }
-
 }
